requirements.txt
python-dotenv
openai==1.58.1